+++ Cargo.toml
@@ -27,6 +27,7 @@ gio = "0.18.4"
 glib = { version = "0.18.5", features = ["v2_78"] }
 gtk4-layer-shell = "0.2.0"
 libhelium = { version = "0.1.0", features = ["gtk_v4_10", "v1_6"] }
+chrono = { version = "0.4.35", features = ["serde"] }
 
 [dependencies.gtk]
 version = "0.7.0"
