--- src/ui/fleet.blp
@@ -1,20 +1,0 @@
-using Gtk 4.0;
-using He 1;
-
-template $Fleet : He.ApplicationWindow {
-    default-height: 25;
-    height-request: 25;
-
-    title: "Fleet Panel";
-
-    Box appbox {
-        orientation: horizontal;
-        spacing: 5;
-
-        Button {
-            label: "Apps";
-            icon-name: "view-app-grid-symbolic";
-        }
-    }
-
-}
