+++ src/widgets/clock/clock.blp
@@ -1,0 +1,10 @@
+using Gtk 4.0;
+using He 1;
+
+template $Clock : Gtk.Box {
+    Box {
+        Label time {
+            label: "00:00:00";
+        }
+    }
+}
