+++ src/widgets/clock/mod.rs
@@ -1,0 +1,8 @@
+use chrono::prelude::*;
+
+pub enum TimeFormat {
+    TwentyFourHour,
+    TwelveHour,
+}
+
+// todo: actually create widget
