+++ src/widgets/mod.rs
@@ -1,0 +1,2 @@
+pub mod clock;
+pub mod fleet;
